EchoBase in lib-builtin-effects
