+++ src/effects/Echo.cpp
@@ -2,8 +2,6 @@
 
   Audacity: A Digital Audio Editor
 
-  Echo.cpp
-
   Dominic Mazzoni
   Vaughan Johnson (dialog)
 
@@ -28,148 +26,12 @@
 #include "AudacityMessageBox.h"
 #include "../widgets/valnum.h"
 
-const EffectParameterMethods& EchoBase::Parameters() const
-{
-   static CapturedParameters<EchoBase,
-      Delay, Decay
-   > parameters;
-   return parameters;
-}
-
-const ComponentInterfaceSymbol EchoBase::Symbol
-{ XO("Echo") };
-
 namespace{ BuiltinEffectsModule::Registration< EffectEcho > reg; }
-
-
-struct EchoBase::Instance
-   : public PerTrackEffect::Instance
-   , public EffectInstanceWithBlockSize
-{
-   Instance(const PerTrackEffect& effect)
-      : PerTrackEffect::Instance{ effect }
-   {}
-
-   bool ProcessInitialize(EffectSettings& settings, double sampleRate,
-      ChannelNames chanMap) override;
-
-   size_t ProcessBlock(EffectSettings& settings,
-      const float* const* inBlock, float* const* outBlock, size_t blockLen)  override;
-
-   bool ProcessFinalize() noexcept override;
-
-   unsigned GetAudioOutCount() const override
-   {
-      return 1;
-   }
-
-   unsigned GetAudioInCount() const override
-   {
-      return 1;
-   }
-
-   Floats history;
-   size_t histPos;
-   size_t histLen;
-};
-
-
 
 std::shared_ptr<EffectInstance> EffectEcho::MakeInstance() const
 {
    return std::make_shared<Instance>(*this);
 }
-
-
-
-
-EchoBase::EchoBase()
-{
-   SetLinearEffectFlag(true);
-}
-
-EchoBase::~EchoBase()
-{
-}
-
-// ComponentInterface implementation
-
-ComponentInterfaceSymbol EchoBase::GetSymbol() const
-{
-   return Symbol;
-}
-
-TranslatableString EchoBase::GetDescription() const
-{
-   return XO("Repeats the selected audio again and again");
-}
-
-ManualPageID EchoBase::ManualPage() const
-{
-   return L"Echo";
-}
-
-// EffectDefinitionInterface implementation
-
-EffectType EchoBase::GetType() const
-{
-   return EffectTypeProcess;
-}
-
-bool EchoBase::Instance::ProcessInitialize(
-   EffectSettings& settings, double sampleRate, ChannelNames)
-{
-   auto& echoSettings = GetSettings(settings);
-   if (echoSettings.delay == 0.0)
-      return false;
-
-   histPos = 0;
-   auto requestedHistLen = (sampleCount) (sampleRate * echoSettings.delay);
-
-   // Guard against extreme delay values input by the user
-   try {
-      // Guard against huge delay values from the user.
-      // Don't violate the assertion in as_size_t
-      if (requestedHistLen !=
-            (histLen = static_cast<size_t>(requestedHistLen.as_long_long())))
-         throw std::bad_alloc{};
-      history.reinit(histLen, true);
-   }
-   catch ( const std::bad_alloc& ) {
-      EffectUIServices::DoMessageBox(mProcessor,
-         XO("Requested value exceeds memory capacity."));
-      return false;
-   }
-
-   return history != NULL;
-}
-
-bool EchoBase::Instance::ProcessFinalize() noexcept
-{
-   return true;
-}
-
-size_t EchoBase::Instance::ProcessBlock(EffectSettings& settings,
-   const float *const *inBlock, float *const *outBlock, size_t blockLen)
-{
-   auto& echoSettings = GetSettings(settings);
-
-   const float *ibuf = inBlock[0];
-   float *obuf = outBlock[0];
-
-   for (decltype(blockLen) i = 0; i < blockLen; i++, histPos++)
-   {
-      if (histPos == histLen)
-      {
-         histPos = 0;
-      }
-      history[histPos] = obuf[i] = ibuf[i] + history[histPos] * echoSettings.decay;
-   }
-
-   return blockLen;
-}
-
-
 
 struct EffectEcho::Editor
    : EffectEditor
